pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.14",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "polars>=1.31.0",
    "pyarrow>=20.0.0",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-jwt>=4.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
